refactor: extract spoil anyway action
